Refactor ProductClient to use pydantic types
