+++ src/pypaystack2/paystack.py
@@ -14,7 +14,6 @@ from pypaystack2.sub_clients.async_clients.payment_requests import (
     AsyncPaymentRequestClient,
 )
 from pypaystack2.sub_clients.async_clients.plans import AsyncPlanClient
-from pypaystack2.sub_clients.products import AsyncProductClient, ProductClient
 from pypaystack2.sub_clients.refunds import AsyncRefundClient, RefundClient
 from pypaystack2.sub_clients.settlements import AsyncSettlementClient, SettlementClient
 from pypaystack2.sub_clients.splits import (
@@ -55,6 +54,10 @@ from pypaystack2.sub_clients.sync_clients.payment_requests import (
     PaymentRequestClient,
 )
 from pypaystack2.sub_clients.sync_clients.plans import PlanClient
+from pypaystack2.sub_clients.sync_clients.products import (
+    AsyncProductClient,
+    ProductClient,
+)
 from pypaystack2.sub_clients.terminals import TerminalClient, AsyncTerminalClient
 from pypaystack2.sub_clients.transactions import (
     AsyncTransactionClient,
